add component to calculate feature importance (#451)
* add component to calculate feature importance

* add log to mltable with correct format

* fix code health

* fix issues with writing to mltable
