+++ responsibleai/model_monitor/components/feature_importance/spec.yaml
@@ -1,0 +1,20 @@
+$schema: http://azureml/sdk-2-0/CommandComponent.json
+name: feature_importance
+version: 0.0.1
+display_name: Feature importance
+is_deterministic: False
+type: command
+description: Feature importance using model monitoring.
+inputs:
+  baseline_data:
+    type: mltable
+outputs:
+  feature_importance_data:
+    type: mltable
+code: ./
+environment: azureml://registries/azureml/environments/responsibleai-ubuntu20.04-py38-cpu/versions/10
+resources:
+  instance_count: 1
+command: python compute_feature_importance.py
+  --baseline_data ${{inputs.baseline_data}}
+  --feature_importance_data ${{outputs.feature_importance_data}}
