+++ responsibleai/model_monitor/components/src/io_utils.py
@@ -4,7 +4,9 @@
 """This file contains utilities to read data."""
 
 import mltable
+import os
 import pandas as pd
+import uuid
 
 
 def load_mltable_to_df(mltable_path) -> pd.DataFrame:
@@ -17,3 +19,17 @@ def load_mltable_to_df(mltable_path) -> pd.DataFrame:
 
     """
     return mltable.load(mltable_path).to_pandas_dataframe()
+
+
+def save_df_as_mltable(df: pd.DataFrame, folder_path: str):
+    """Write a dataframe to MLTable file.
+
+    Args:
+        df: The DataFrame to save.
+        folder_path: The folder path in Azure blob store to save the MLTable to.
+
+    """
+    filename = f'./{uuid.uuid4()}.parquet'
+
+    df.to_parquet(os.path.join(folder_path, filename), index=False)
+    mltable.from_parquet_files(paths=[{'file': filename}]).save(path=folder_path)
